requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.0
matplotlib==3.8.4
plotly==5.22.0
requests==2.31.0
tqdm==4.66.1
astroquery==0.4.7
PyYAML==6.0.1
